+++ Frameworks/PWEngagement.xcframework/ios-arm64_x86_64-simulator/PWEngagement.framework/Headers/PWEngagement-Swift.h
@@ -260,16 +260,17 @@ using UInt = size_t;
 SWIFT_CLASS("_TtC12PWEngagement7Message")
 @interface Message : NSObject <NSCopying>
 @property (nonatomic, readonly, copy) NSString * _Nullable identifier;
-@property (nonatomic, readonly, copy) NSString * _Nonnull campaignID;
+@property (nonatomic, readonly) NSInteger campaignIdentifier;
 @property (nonatomic, readonly, copy) NSString * _Nonnull campaignType;
 @property (nonatomic, readonly, copy) NSDictionary<NSString *, NSString *> * _Nullable metaData;
 @property (nonatomic, readonly, copy) NSString * _Nullable alertTitle;
 @property (nonatomic, readonly, copy) NSString * _Nullable alertBody;
 @property (nonatomic, readonly, copy) NSString * _Nullable promotionTitle;
 @property (nonatomic, readonly, copy) NSString * _Nullable promotionBody;
+@property (nonatomic, readonly, copy) NSDate * _Nullable promotionExpirationDate;
 @property (nonatomic, readonly) BOOL isRead;
 @property (nonatomic, readonly, copy) NSDate * _Nullable timestamp;
-- (nonnull instancetype)initWithIdentifier:(NSString * _Nullable)identifier campaignID:(NSString * _Nonnull)campaignID alertTitle:(NSString * _Nullable)alertTitle alertBody:(NSString * _Nullable)alertBody promotionTitle:(NSString * _Nullable)promotionTitle promotionBody:(NSString * _Nullable)promotionBody metaData:(NSDictionary<NSString *, NSString *> * _Nullable)metaData campaignType:(NSString * _Nonnull)campaignType isRead:(BOOL)isRead timestamp:(NSDate * _Nullable)timestamp OBJC_DESIGNATED_INITIALIZER;
+- (nonnull instancetype)initWithIdentifier:(NSString * _Nullable)identifier campaignIdentifier:(NSInteger)campaignIdentifier alertTitle:(NSString * _Nullable)alertTitle alertBody:(NSString * _Nullable)alertBody promotionTitle:(NSString * _Nullable)promotionTitle promotionBody:(NSString * _Nullable)promotionBody promotionExpirationDate:(NSDate * _Nullable)promotionExpirationDate metaData:(NSDictionary<NSString *, NSString *> * _Nullable)metaData campaignType:(NSString * _Nonnull)campaignType isRead:(BOOL)isRead timestamp:(NSDate * _Nullable)timestamp OBJC_DESIGNATED_INITIALIZER;
 + (Message * _Nullable)unpackWithDict:(NSDictionary<NSString *, id> * _Nonnull)dict SWIFT_WARN_UNUSED_RESULT;
 - (id _Nonnull)copyWithZone:(struct _NSZone * _Nullable)zone SWIFT_WARN_UNUSED_RESULT;
 - (nonnull instancetype)init SWIFT_UNAVAILABLE;
@@ -277,10 +278,33 @@ SWIFT_CLASS("_TtC12PWEngagement7Message")
 @end
 
 
-@interface PWEngagement (SWIFT_EXTENSION(PWEngagement))
-+ (void)postMessageEventWithMessageIdentifier:(NSString * _Nonnull)messageIdentifier completion:(void (^ _Nonnull)(NSError * _Nullable))completion;
+@interface Message (SWIFT_EXTENSION(PWEngagement))
++ (NSDate * _Nullable)promotionExpirationDateFromString:(NSString * _Nullable)dateString SWIFT_WARN_UNUSED_RESULT;
 @end
 
+
+SWIFT_CLASS("_TtC12PWEngagement12MessageEvent")
+@interface MessageEvent : NSObject <NSCopying>
+@property (nonatomic, readonly, copy) NSString * _Nonnull identifier;
+@property (nonatomic, readonly, copy) NSString * _Nonnull deviceIdentifier;
+@property (nonatomic, readonly) NSInteger appIdentifier;
+@property (nonatomic, readonly) NSInteger campaignIdentifier;
+@property (nonatomic, readonly) BOOL isRead;
+@property (nonatomic, readonly) NSInteger sentTime;
+- (id _Nonnull)copyWithZone:(struct _NSZone * _Nullable)zone SWIFT_WARN_UNUSED_RESULT;
+- (nonnull instancetype)init SWIFT_UNAVAILABLE;
++ (nonnull instancetype)new SWIFT_UNAVAILABLE_MSG("-init is unavailable");
+@end
+
+
+@interface PWEngagement (SWIFT_EXTENSION(PWEngagement))
++ (void)postMessageEventWithCampaignIdentifier:(NSString * _Nonnull)campaignIdentifier completion:(void (^ _Nonnull)(NSString * _Nullable, NSError * _Nullable))completion;
+@end
+
+
+@interface PWEngagement (SWIFT_EXTENSION(PWEngagement))
++ (void)fetchMessageWithIdentifier:(NSString * _Nonnull)identifier completion:(void (^ _Nonnull)(Message * _Nullable, NSError * _Nullable))completion;
+@end
 
 @class UNNotification;
 @class UNNotificationResponse;
@@ -563,16 +587,17 @@ using UInt = size_t;
 SWIFT_CLASS("_TtC12PWEngagement7Message")
 @interface Message : NSObject <NSCopying>
 @property (nonatomic, readonly, copy) NSString * _Nullable identifier;
-@property (nonatomic, readonly, copy) NSString * _Nonnull campaignID;
+@property (nonatomic, readonly) NSInteger campaignIdentifier;
 @property (nonatomic, readonly, copy) NSString * _Nonnull campaignType;
 @property (nonatomic, readonly, copy) NSDictionary<NSString *, NSString *> * _Nullable metaData;
 @property (nonatomic, readonly, copy) NSString * _Nullable alertTitle;
 @property (nonatomic, readonly, copy) NSString * _Nullable alertBody;
 @property (nonatomic, readonly, copy) NSString * _Nullable promotionTitle;
 @property (nonatomic, readonly, copy) NSString * _Nullable promotionBody;
+@property (nonatomic, readonly, copy) NSDate * _Nullable promotionExpirationDate;
 @property (nonatomic, readonly) BOOL isRead;
 @property (nonatomic, readonly, copy) NSDate * _Nullable timestamp;
-- (nonnull instancetype)initWithIdentifier:(NSString * _Nullable)identifier campaignID:(NSString * _Nonnull)campaignID alertTitle:(NSString * _Nullable)alertTitle alertBody:(NSString * _Nullable)alertBody promotionTitle:(NSString * _Nullable)promotionTitle promotionBody:(NSString * _Nullable)promotionBody metaData:(NSDictionary<NSString *, NSString *> * _Nullable)metaData campaignType:(NSString * _Nonnull)campaignType isRead:(BOOL)isRead timestamp:(NSDate * _Nullable)timestamp OBJC_DESIGNATED_INITIALIZER;
+- (nonnull instancetype)initWithIdentifier:(NSString * _Nullable)identifier campaignIdentifier:(NSInteger)campaignIdentifier alertTitle:(NSString * _Nullable)alertTitle alertBody:(NSString * _Nullable)alertBody promotionTitle:(NSString * _Nullable)promotionTitle promotionBody:(NSString * _Nullable)promotionBody promotionExpirationDate:(NSDate * _Nullable)promotionExpirationDate metaData:(NSDictionary<NSString *, NSString *> * _Nullable)metaData campaignType:(NSString * _Nonnull)campaignType isRead:(BOOL)isRead timestamp:(NSDate * _Nullable)timestamp OBJC_DESIGNATED_INITIALIZER;
 + (Message * _Nullable)unpackWithDict:(NSDictionary<NSString *, id> * _Nonnull)dict SWIFT_WARN_UNUSED_RESULT;
 - (id _Nonnull)copyWithZone:(struct _NSZone * _Nullable)zone SWIFT_WARN_UNUSED_RESULT;
 - (nonnull instancetype)init SWIFT_UNAVAILABLE;
@@ -580,10 +605,33 @@ SWIFT_CLASS("_TtC12PWEngagement7Message")
 @end
 
 
-@interface PWEngagement (SWIFT_EXTENSION(PWEngagement))
-+ (void)postMessageEventWithMessageIdentifier:(NSString * _Nonnull)messageIdentifier completion:(void (^ _Nonnull)(NSError * _Nullable))completion;
+@interface Message (SWIFT_EXTENSION(PWEngagement))
++ (NSDate * _Nullable)promotionExpirationDateFromString:(NSString * _Nullable)dateString SWIFT_WARN_UNUSED_RESULT;
 @end
 
+
+SWIFT_CLASS("_TtC12PWEngagement12MessageEvent")
+@interface MessageEvent : NSObject <NSCopying>
+@property (nonatomic, readonly, copy) NSString * _Nonnull identifier;
+@property (nonatomic, readonly, copy) NSString * _Nonnull deviceIdentifier;
+@property (nonatomic, readonly) NSInteger appIdentifier;
+@property (nonatomic, readonly) NSInteger campaignIdentifier;
+@property (nonatomic, readonly) BOOL isRead;
+@property (nonatomic, readonly) NSInteger sentTime;
+- (id _Nonnull)copyWithZone:(struct _NSZone * _Nullable)zone SWIFT_WARN_UNUSED_RESULT;
+- (nonnull instancetype)init SWIFT_UNAVAILABLE;
++ (nonnull instancetype)new SWIFT_UNAVAILABLE_MSG("-init is unavailable");
+@end
+
+
+@interface PWEngagement (SWIFT_EXTENSION(PWEngagement))
++ (void)postMessageEventWithCampaignIdentifier:(NSString * _Nonnull)campaignIdentifier completion:(void (^ _Nonnull)(NSString * _Nullable, NSError * _Nullable))completion;
+@end
+
+
+@interface PWEngagement (SWIFT_EXTENSION(PWEngagement))
++ (void)fetchMessageWithIdentifier:(NSString * _Nonnull)identifier completion:(void (^ _Nonnull)(Message * _Nullable, NSError * _Nullable))completion;
+@end
 
 @class UNNotification;
 @class UNNotificationResponse;